Create Separate ViewModel
